Modify Makefile to compile deauth and PAM
During testing, there seems to be something wrong with the rfcomm server. Need to debug why the phone is not sending any messages
